requirements.txt
numpy
pysoundfile
pyTelegramBotAPI
